Change input type from int to float for dimensions

+++ area-of-rectangle.py
@@ -1,4 +1,4 @@
-length = int(input("Enter the length of the rectangle: "))
-breadth = int(input("Enter the width of the rectangle: "))
+length = float(input("Enter the length of the rectangle: "))
+breadth = float(input("Enter the width of the rectangle: "))
 area = length * breadth
 print(area)
